requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apispec==6.8.2
async-substrate-interface==1.4.1
async-timeout==5.0.1
asyncstdlib==3.13.1
attrs==25.3.0
backoff==2.2.1
backports-datetime-fromisoformat==2.0.3
base58==2.1.1
bittensor==9.8.3
bittensor-cli==9.8.7
bittensor-drand==0.5.1
bittensor-wallet==3.1.0
blinker==1.9.0
bt-decode==0.6.0
cachelib==0.13.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
decorator==5.2.1
exceptiongroup==1.3.0
fastapi==0.110.3
Flask==3.1.1
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
flask-smorest==0.46.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.1
msgpack-numpy-opentensor==0.5.0
multidict==6.6.3
munch==2.5.0
narwhals==1.47.1
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==2.2.6
packaging==25.0
plotille==5.0.0
plotly==6.2.0
propcache==0.3.2
psycopg2-binary==2.9.10
py==1.11.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-statemachine==2.5.0
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
retry==0.9.2
rich==14.0.0
scalecodec==1.2.11
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.37.2
toml==0.10.0
tomli==2.2.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
webargs==8.7.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
